tools/signal_mcp/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "signal-mcp"
version = "0.1.0"
description = "Signal messaging MCP server backed by signal-cli REST API"
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.1",
]

[project.scripts]
signal-mcp = "signal_mcp.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["signal_mcp*"]
